+++ src/Screens/Orders/types.ts
@@ -1,5 +1,6 @@
 export interface ComponentType {
-    openOrders: boolean;
-    deliveredOrders: boolean;
-    cancelledOrders: boolean;
-  }
+  allOrders: boolean;
+  openOrders: boolean;
+  deliveredOrders: boolean;
+  cancelledOrders: boolean;
+}
