+++ src/redux/rootReducer.ts
@@ -1,13 +1,14 @@
 import {combineReducers} from '@reduxjs/toolkit';
 import userReducer from './slices/auth/userSlice';
-import orderReducer from './slices/order/orderSlice';
+import cartReducer from './slices/cart/cartSlice';
 import guestReducer from './slices/auth/guestSlice';
-
+import orderReducer from './slices/order/orderSlice';
 
 const rootReducer = combineReducers({
   user: userReducer,
   guest: guestReducer,
-  order:orderReducer
+  cart: cartReducer,
+  order: orderReducer,
 });
 
 export type RootState = ReturnType<typeof rootReducer>;
