+++ src/redux/slices/order/orderServices.ts
@@ -3,129 +3,46 @@ import axios from 'axios';
 import {apiUrl} from '../../../config';
 import {apiCall} from '../../../utils';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import {
-  ProductListResponse,
-  ProductFeeResponse,
-  CartItemResponsePayload,
-  FeesResponse,
-  StateResponse,
-} from './types';
-
-export interface FeeTypeData {
-  filter: {
-    fields: {
-      name: string;
-      operator: string;
-      value: string;
-    }[];
-    operator: string;
-  };
-  paging: {
-    index: number;
-    size: number;
-  };
-}
-
-export interface CartData {
-  cost: number;
-  product_id: string;
-  quantity: number;
-  weight: number;
-}
-export interface CartItemQuantityData {
-  cart_item_id: string;
-  quantity: number;
-}
-
-export interface FeesData {
-  filter: {
-    fields: {
-      name: string;
-      operator: string;
-      value: string;
-    }[];
-    operator: string;
-  };
-  paging: {
-    index: number;
-    size: number;
-  };
-}
 
 export interface PaginationData {
-  // filter: {
-  //   fields: {
-  //     name: string;
-  //     operator: string;
-  //     value: string;
-  //   }[];
-  //   operator: string;
-  // };
+  //   filter: {
+  //     fields: [
+  //       {
+  //         keys: ['string'];
+  //         name: 'string';
+  //         operator: 'beginsWith';
+  //         value: 'string';
+  //       },
+  //     ];
+  //     operator: 'and';
+  //   };
   paging: {
     index: number;
     size: number;
   };
+  //   sorting: {
+  //     column: 'string';
+  //     direction: 'desc';
+  //   };
 }
 
-export interface DeliveryFeeData {
-  filter: {
-    fields: [
-      {
-        name: string,
-        operator: 'isEqualTo',
-        value: string,
-      },
-      {
-        name: string,
-        operator: 'isEqualTo',
-        value: string,
-      },
-    ],
-    operator: string;
-  };
-  paging: {
-    index: number;
-    size: number;
-  };
-}
-
-export interface CheckoutData {
-  cart_id: string;
-  delivery_details: {
-    address: {
-      city: string;
-      closest_landmark: string;
-      coordinate: {
-        latitude: number;
-        longitude: number;
-      };
-      full_address: string;
-      lga: string;
-      state: string;
-      verified: boolean;
-    };
-    email: string;
-    name: string;
-    phone: string;
-  };
-  delivery_fee: number;
-  payment_method: string;
-  service_fee: number;
-  total_fee: number;
-}
-
-export const productList = createAsyncThunk(
-  'order/productList',
-  async (productList: PaginationData, {rejectWithValue}) => {
+export const triggerOrderList = createAsyncThunk(
+  'order/order_list',
+  async (data: PaginationData, {rejectWithValue}) => {
     try {
-      const url = apiUrl.productList;
+      const url = apiUrl.getOrders;
       const method = 'put';
       const token = await AsyncStorage.getItem('userToken');
       const headers = {
         Authorization: `Bearer ${token}`,
       };
-      const response = await apiCall(url, method, productList, headers);
-      return response as ProductListResponse;
+      const response = (await apiCall(
+        url,
+        method,
+        data,
+        headers,
+      )) as unknown as Record<string, any>;
+      return response.data;
     } catch (error) {
       if (axios.isAxiosError(error)) {
         return rejectWithValue(error.response?.data);
@@ -138,213 +55,21 @@ export const productList = createAsyncThunk(
   },
 );
 
-export const productFee = createAsyncThunk(
-  'order/productFee',
-  async (feeType: FeeTypeData, {rejectWithValue}) => {
+export const triggerOrderDetails = createAsyncThunk(
+  'order/order_details',
+  async (orderid, {rejectWithValue}) => {
     try {
-      const url = apiUrl.feesType;
-      const method = 'put';
-      const token = await AsyncStorage.getItem('userToken');
-      const headers = {
-        Authorization: `Bearer ${token}`,
-      };
-      const response = await apiCall(url, method, feeType, headers);
-      return response as ProductFeeResponse;
-    } catch (error) {
-      if (axios.isAxiosError(error)) {
-        return rejectWithValue(error.response?.data);
-      } else if (error instanceof Error) {
-        return rejectWithValue(error.message);
-      } else {
-        throw error;
-      }
-    }
-  },
-);
-
-export const addTocart = createAsyncThunk(
-  'order/addTocart',
-  async (cart: CartData, {rejectWithValue}) => {
-    try {
-      const url = apiUrl.cartAdd;
-      const method = 'post';
-      const token = await AsyncStorage.getItem('userToken');
-      const headers = {
-        Authorization: `Bearer ${token}`,
-      };
-      const response = await apiCall(url, method, cart, headers);
-      return response as CartItemResponsePayload;
-    } catch (error) {
-      if (axios.isAxiosError(error)) {
-        return rejectWithValue(error.response?.data);
-      } else if (error instanceof Error) {
-        return rejectWithValue(error.message);
-      } else {
-        throw error;
-      }
-    }
-  },
-);
-
-export const updateCartItemQuantity = createAsyncThunk(
-  'order/updateCartItemQuantity',
-  async (cartItemQuantity: CartItemQuantityData, {rejectWithValue}) => {
-    try {
-      const url = apiUrl.cartItemQuantity;
-      const method = 'put';
-      const token = await AsyncStorage.getItem('userToken');
-      const headers = {
-        Authorization: `Bearer ${token}`,
-      };
-      const response = await apiCall(url, method, cartItemQuantity, headers);
-      return response as Record<string, Record<string, string> | string>;
-    } catch (error) {
-      if (axios.isAxiosError(error)) {
-        return rejectWithValue(error.response?.data);
-      } else if (error instanceof Error) {
-        return rejectWithValue(error.message);
-      } else {
-        throw error;
-      }
-    }
-  },
-);
-
-export const serviceFee = createAsyncThunk(
-  'order/serviceFee',
-  async (fee: FeesData, {rejectWithValue}) => {
-    try {
-      const url = apiUrl.feesType;
-      const method = 'put';
-      const token = await AsyncStorage.getItem('userToken');
-      const headers = {
-        Authorization: `Bearer ${token}`,
-      };
-      const response = await apiCall(url, method, fee, headers);
-      return response as FeesResponse;
-    } catch (error) {
-      if (axios.isAxiosError(error)) {
-        return rejectWithValue(error.response?.data);
-      } else if (error instanceof Error) {
-        return rejectWithValue(error.message);
-      } else {
-        throw error;
-      }
-    }
-  },
-);
-
-export const triggerCartList = createAsyncThunk(
-  'order/cartList',
-  async (cartList: PaginationData, {rejectWithValue}) => {
-    try {
-      const url = apiUrl.listCart;
-      const method = 'put';
-      const token = await AsyncStorage.getItem('userToken');
-      const headers = {
-        Authorization: `Bearer ${token}`,
-      };
-      const response = await apiCall(url, method, cartList, headers);
-      return response as CartItemResponsePayload;
-    } catch (error) {
-      if (axios.isAxiosError(error)) {
-        return rejectWithValue(error.response?.data);
-      } else if (error instanceof Error) {
-        return rejectWithValue(error.message);
-      } else {
-        throw error;
-      }
-    }
-  },
-);
-
-export const getState = createAsyncThunk(
-  'order/getState',
-  async (_, {rejectWithValue}) => {
-    try {
-      const url = apiUrl.getStates;
+      const url = apiUrl.getOrders + orderid;
       const method = 'get';
       const token = await AsyncStorage.getItem('userToken');
       const headers = {
         Authorization: `Bearer ${token}`,
       };
-      const response = await apiCall(url, method, undefined, headers);
-      return response as StateResponse;
-    } catch (error) {
-      if (axios.isAxiosError(error)) {
-        return rejectWithValue(error.response?.data);
-      } else if (error instanceof Error) {
-        return rejectWithValue(error.message);
-      } else {
-        throw error;
-      }
-    }
-  },
-);
-
-export const deliveryFee = createAsyncThunk(
-  'order/deliveryFee',
-  async (deliveryFeePayload: DeliveryFeeData, {rejectWithValue}) => {
-    try {
-      const url = apiUrl.feesType;
-      const method = 'put';
-      const token = await AsyncStorage.getItem('userToken');
-      const headers = {
-        Authorization: `Bearer ${token}`,
-      };
-      const response = await apiCall(url, method, deliveryFeePayload, headers);
+      const response = await apiCall(url, method, {}, headers);
       return response;
     } catch (error) {
       if (axios.isAxiosError(error)) {
         return rejectWithValue(error.response?.data);
-      } else if (error instanceof Error) {
-        return rejectWithValue(error.message);
-      } else {
-        throw error;
-      }
-    }
-  },
-);
-
-export const triggerDeleteCartItem = createAsyncThunk(
-  'order/triggerDeleteCartItem',
-  async (id: string, {rejectWithValue}) => {
-    try {
-      const url = `${apiUrl.deleteCartItem}/${id}`;
-      const method = 'delete';
-      const token = await AsyncStorage.getItem('userToken');
-      const headers = {
-        Authorization: `Bearer ${token}`,
-      };
-      const response = await apiCall(url, method, id, headers);
-      return response as Record<string, string>;
-    } catch (error) {
-      if (axios.isAxiosError(error)) {
-        return rejectWithValue(error.response?.data);
-      } else if (error instanceof Error) {
-        return rejectWithValue(error.message);
-      } else {
-        throw error;
-      }
-    }
-  },
-);
-
-export const triggerCheckout = createAsyncThunk(
-  'order/triggerCheckout',
-  async (checkout: CheckoutData, {rejectWithValue}) => {
-    try {
-      const url = apiUrl.checkout;
-      const method = 'post';
-      const token = await AsyncStorage.getItem('userToken');
-      const headers = {
-        Authorization: `Bearer ${token}`,
-      };
-      const response = await apiCall(url, method, checkout, headers);
-      return response as any
-    } catch (error) {
-      if (axios.isAxiosError(error)) {
-        return rejectWithValue(error.response || error.response);
       } else if (error instanceof Error) {
         return rejectWithValue(error.message);
       } else {
